Move the TimeProvider to the Symfony app
The SP Metadata generator will be using this time provider, moving it to
the Symfony Enginblock codebase is now required.
